Fix Issue #290
- Listeners fail to attach if script fails at any point.
- Scripts fail if jQuery fails to find elements
- Elements aren't found if they aren't rendered for auth reasons
- Fix is to skip listener attachments if toolbox isn't rendered

+++ app/assets/javascripts/channels/avatars.coffee
@@ -109,19 +109,19 @@ nameToggle = (data) ->
   App.drawFrame()
 
 document.addEventListener 'turbolinks:load', (e) ->
+  if document.querySelector '#toolbox' != null
+    document.querySelectorAll('.avatar-selection').forEach (elem) ->
+      elem.addEventListener 'mouseup', (e) ->
+        App.avatar.hold this.dataset.avatarId
 
-  document.querySelectorAll('.avatar-selection').forEach (elem) ->
-    elem.addEventListener 'mouseup', (e) ->
-      App.avatar.hold this.dataset.avatarId
+    document.querySelector('#avatarSlider').addEventListener 'change', (e) ->
+      App.avatar.size document.querySelector('#avatarSlider').value
 
-  document.querySelector('#avatarSlider').addEventListener 'change', (e) ->
-    App.avatar.size document.querySelector('#avatarSlider').value
+    document.querySelector('#dropAvatarButton').addEventListener 'mouseup', (e) ->
+      App.avatar.drop()
 
-  document.querySelector('#dropAvatarButton').addEventListener 'mouseup', (e) ->
-    App.avatar.drop()
-
-  document.querySelector('#avatarName').addEventListener 'mouseup', (e) ->
-    App.avatar.nameToggle()
+    document.querySelector('#avatarName').addEventListener 'mouseup', (e) ->
+      App.avatar.nameToggle()
 
   document.querySelector('#canvas').addEventListener 'mouseup', (e) ->
     App.avatar.place e.x, e.y, document.querySelector('#avatarSlider').value, document.querySelector('.avatar-selection').avatarName
